Add a --config-file option.
If a --config-file is explicitly given, then the file MUST exist and be
readable or the server aborts.  Otherwise try to open a default file at
/etc/commissaire/commissaire.conf, but tolerate it not existing.

Note, this also simplifies running Commissaire from a container.  Just
create a commissaire.conf file and mount it under /etc/commissaire in
the container.

Closes: #110
Approved by: ashcrow