contrib/spec: Add bcond for tests (CentOS 7 work)
I was looking at adding commissaire to the Project Atomic rdgo,
but we don't presently have flake8 or mock.

One of the things I'm pushing for as well with the rdgo model
is to always run checks outside of the build as installed tests.
(Though that would apply more to the mock aspect, flake8 is
 kind of like imposing a compiler on Python, so it's more build time.)

Closes: #112
Approved by: ashcrow