doc: Clarified cloud-init integration.
During some UAT testing it became clear that the cloud-init information
was technically correct, but didn't state what the cloud-config.txt
should look like. The user used the default example from cloud-init
which caused his cloud-provider to reject the file due to size. This
change adds an example cloud-config.txt, adds it to the documentation,
calls out that nothing special is needed in the cloud-config.txt, and
attempts to do a few minor clarifications.

Closes: #128
Approved by: cooktheryan