Add a 'store-manager-clone' bus method.
Returns a clone of the StoreHandlerManager instance in StorePlugin.
Typically passed as an argument to a multiprocessing.Process object.

Note the investigator process takes a unique StoreHandlerManager
instance for each job.  This was done to avoid InvestigatorPlugin
racing with StorePlugin during initialization.  Plugins shouldn't
depend on other plugins.

Closes: #155
Approved by: ashcrow