bug: Hosts without a cluster now gain the default network.
When a Host is not part of a cluster it doesn't have a network
associated with it. This was causing a problem during bootstrapping. Now
a host outside of any cluster will associate with the default network.

This change also makes sure that cluster_data is usable before
associating with a Container Manager.